emtech_backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-allauth==0.65.10
dj-rest-auth==7.0.1
python-decouple==3.8
dj-database-url==2.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0 